+++ gazebo/rendering/JointVisual.hh
@@ -43,10 +43,29 @@ namespace gazebo
       /// \brief Destructor
       public: virtual ~JointVisual();
 
-      /// \brief Load the visual based on a message
+      /// \brief Load the joint visual based on a message
       /// \param[in] _msg Joint message
       public: void Load(ConstJointPtr &_msg);
       using Visual::Load;
+
+      /// \internal
+      /// \brief Load the joint visual based on a message and an offset pose
+      /// This is currently used internally for creating a second visual for
+      /// joint types that have more than 1 axis.
+      /// \param[in] _msg Joint message
+      /// \param[in] _pose Pose of the joint visual in world coordinates.
+      public: void Load(ConstJointPtr &_msg, const math::Pose &_worldPose);
+
+      /// \brief Create an axis and attach it to the joint visual.
+      /// \param[in] _axis Axis vector
+      /// \param[in] _useParentFrame True to use parent frame instead of the
+      /// joint frame.
+      /// \param[in] _type Type of axis.
+      public: void CreateAxis(const math::Vector3 &_axis, bool _useParentFrame,
+          msgs::Joint::Type _type);
+
+      // Documentation Inherited.
+      public: void SetVisible(bool _visible, bool _cascade = true);
     };
     /// \}
   }