+++ gazebo/rendering/ArrowVisual.hh
@@ -44,8 +44,17 @@ namespace gazebo
       /// \brief Load the visual with default parameters
       public: virtual void Load();
 
-      /// \brief Show the rotation of the visual
-      public: void ShowRotation();
+      /// \brief Show the rotation of the arrow visual.
+      /// \param[in] _show True to show the arrow rotation.
+      public: void ShowRotation(bool _show);
+
+      /// \brief Show the shaft of the arrow visual.
+      /// \param[in] _show True to show the arrow shaft.
+      public: void ShowShaft(bool _show);
+
+      /// \brief Show the head of the arrow visual.
+      /// \param[in] _show True to show the arrow head.
+      public: void ShowHead(bool _show);
     };
     /// \}
   }